% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qdaClassification.R
\name{qdaClassification}
\alias{qdaClassification}
\title{qdaClassification}
\usage{
qdaClassification(normal, data)
}
\arguments{
\item{normal}{A list with arguments mean, covaruance and weight.}

\item{data}{Data frame or matrix on which to perform qda.}
}
\value{
A score for each point.
}
\description{
Gives quadratic discriminant scores to the points in data for a multivariate normal.
}
\examples{
data.qda = cbind(rnorm(50), rnorm(50))
exp(qdaClassification(list(mean = c(0,0), cov = diag(1,2), weight = 1), data.qda))

}
