% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2dist.R
\name{w2dist}
\alias{w2dist}
\title{w2dist}
\usage{
w2dist(P, Q)
}
\arguments{
\item{P}{A multivariate normal distribution given as a list with arguments mean and cov.}

\item{Q}{A multivariate normal distribution given as a list with arguments mean and cov.}
}
\value{
A double giving the 2-Wasserstein distance between the two distributions.
}
\description{
The 2-Wasserstein distance between two multivariate normal distributions
}
\examples{
P <- list(mean = c(1, 1), cov = diag(1, 2))
Q <- list(mean = c(0, 0), cov = 1.1*diag(1, 2))
Q <- list(mean = c(0, 0), cov = 1.1*diag(1, 2))
w2dist(P, Q)

}
